***Changed Sharer and Leaderboard use cases so the display of the Leaderboard (with emojis) is under Leaderboard
