+++ src/main/java/share_use_case/ShareController.java
@@ -1,10 +1,12 @@
 package share_use_case;
 
+import LeaderboardUseCase.LeaderboardController;
+
 public class ShareController {
 
     public StringBuilder whatistheRankings() {
 
-        return Sharer.getRanking();
+        return LeaderboardController.getRanking();
     }
 
     public StringBuilder whatisthefarminEmojis() {
